Separated out authentication logic from main application.
